Remove empty intersections
