+++ testdata/jennies/rawtypes/reference_of_reference/PHPRawTypes/src/ReferenceOfReference/OtherStruct.php
@@ -2,4 +2,37 @@
 
 namespace Grafana\Foundation\ReferenceOfReference;
 
-class OtherStruct extends \Grafana\Foundation\ReferenceOfReference\AnotherStruct {}
+class OtherStruct implements \JsonSerializable
+{
+    public string $a;
+
+    /**
+     * @param string|null $a
+     */
+    public function __construct(?string $a = null)
+    {
+        $this->a = $a ?: "";
+    }
+
+    /**
+     * @param array<string, mixed> $inputData
+     */
+    public static function fromArray(array $inputData): self
+    {
+        /** @var array{a?: string} $inputData */
+        $data = $inputData;
+        return new self(
+            a: $data["a"] ?? null,
+        );
+    }
+
+    /**
+     * @return mixed
+     */
+    public function jsonSerialize(): mixed
+    {
+        $data = new \stdClass;
+        $data->a = $this->a;
+        return $data;
+    }
+}
