+++ testdata/jennies/rawtypes/refs/PHPRawTypes/src/Refs/RefToSomeStruct.php
@@ -2,4 +2,40 @@
 
 namespace Grafana\Foundation\Refs;
 
-class RefToSomeStruct extends \Grafana\Foundation\Refs\SomeStruct {}
+class RefToSomeStruct implements \JsonSerializable
+{
+    /**
+     * @var mixed
+     */
+    public $fieldAny;
+
+    /**
+     * @param mixed|null $fieldAny
+     */
+    public function __construct( $fieldAny = null)
+    {
+        $this->fieldAny = $fieldAny ?: null;
+    }
+
+    /**
+     * @param array<string, mixed> $inputData
+     */
+    public static function fromArray(array $inputData): self
+    {
+        /** @var array{FieldAny?: mixed} $inputData */
+        $data = $inputData;
+        return new self(
+            fieldAny: $data["FieldAny"] ?? null,
+        );
+    }
+
+    /**
+     * @return mixed
+     */
+    public function jsonSerialize(): mixed
+    {
+        $data = new \stdClass;
+        $data->FieldAny = $this->fieldAny;
+        return $data;
+    }
+}
