--- testdata/jennies/rawtypes/refs/PHPRawTypes/src/Refs/SomeStruct.php
@@ -1,41 +1,0 @@
-<?php
-
-namespace Grafana\Foundation\Refs;
-
-class SomeStruct implements \JsonSerializable
-{
-    /**
-     * @var mixed
-     */
-    public $fieldAny;
-
-    /**
-     * @param mixed|null $fieldAny
-     */
-    public function __construct( $fieldAny = null)
-    {
-        $this->fieldAny = $fieldAny ?: null;
-    }
-
-    /**
-     * @param array<string, mixed> $inputData
-     */
-    public static function fromArray(array $inputData): self
-    {
-        /** @var array{FieldAny?: mixed} $inputData */
-        $data = $inputData;
-        return new self(
-            fieldAny: $data["FieldAny"] ?? null,
-        );
-    }
-
-    /**
-     * @return mixed
-     */
-    public function jsonSerialize(): mixed
-    {
-        $data = new \stdClass;
-        $data->FieldAny = $this->fieldAny;
-        return $data;
-    }
-}
